+++ NEWS.md
@@ -4,6 +4,7 @@
 
 * Added MDR-TB data for 2016, see [here](http://www.who.int/tb/country/data/download/en/) for the dataset. The MDR-TB data is automatically joined to the WHO TB burden data.
 * Aesthetic updates to `map_tb_burden`.
+* Added new `summarise_tb_burden` function for summarising metrics across regions, across custom groups and globally.
 
 ## Package updates